api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0",
    "strawberry-graphql-django>=0.72.0",
    "django-cors-headers>=4.0.0",
]
